Use CRC32 for hashing seed string
This change hashes the user submitted seed value with CRC32.

The reason for this is that calling seed(n), with n being a string and not a numeric value, forces Python to hash the string with the built-in hash() function which creates different results on different operating systems and maybe even different Python versions.

Using a specific yet simple algorithm like CRC32 guarantees that players on Mac OSX, Linux and Windows always get the same three starting items which at the moment is not the case (tested on Linux vs. Windows).